+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## unreleased
 
+- Support aggregation on Glottolog dialect level as well.
 - It turns out that fixing polygons may also create artefacts of type `MultiPoint`.
 - Fixed problem where `geojson.multipolygon_spread` relied on a `Glottolog_Languoid_Level` column
   being present.